Refactor Game.java and Scene.java to support scene management
Game.java and Scene.java has been refactored to support multiple scenes
and switching between them. All gameloop logic has been moved from Scene
to Game. Scene now only represents a collection of game objects and a
map of transitions to other scenes. Game is now responsible for setting
up the game window, handling user input, timing, and drawing to the
screen.
